requirements.txt
Flask
assemblyai
python-dotenv
requests
